--- device_doctor/config.yaml
@@ -1,33 +1,0 @@
-# The device operating system for different bots. Possible values: android, ios.
-# This configuration file tracks the Flutter devicelab hardware status, and 
-# needs to be updated whatever changes from real world.
-flutter-devicelab-linux-1: android
-flutter-devicelab-linux-2: android
-flutter-devicelab-linux-3: android
-flutter-devicelab-linux-4: android
-flutter-devicelab-linux-5: android
-flutter-devicelab-linux-6: android
-flutter-devicelab-linux-7: android
-flutter-devicelab-linux-8: android
-flutter-devicelab-mac-1.mtv.corp.google.com: ios
-flutter-devicelab-mac-2.mtv.corp.google.com: ios
-flutter-devicelab-mac-3.mtv.corp.google.com: ios
-flutter-devicelab-mac-4.mtv.corp.google.com: ios
-flutter-devicelab-mac-5.mtv.corp.google.com: ios
-flutter-devicelab-mac-6.mtv.corp.google.com: ios
-flutter-devicelab-mac-7.mtv.corp.google.com: ios
-flutter-devicelab-mac-8.mtv.corp.google.com: ios
-flutter-devicelab-mac-9.mtv.corp.google.com: ios
-flutter-devicelab-mac-10.mtv.corp.google.com: ios
-flutter-devicelab-mac-11.mtv.corp.google.com: ios
-flutter-devicelab-mac-12.mtv.corp.google.com: ios
-flutter-devicelab-mac-13.mtv.corp.google.com: android
-flutter-devicelab-mac-14.mtv.corp.google.com: ios
-flutter-devicelab-mac-17.mtv.corp.google.com: android
-flutter-devicelab-mac-18.mtv.corp.google.com: android
-flutter-devicelab-mac-19.mtv.corp.google.com: android
-flutter-devicelab-mac-20.mtv.corp.google.com: android
-flutter-devicelab-mac-21.mtv.corp.google.com: android
-flutter-devicelab-mac-22.mtv.corp.google.com: android
-flutter-devicelab-windows-1: android
-flutter-devicelab-windows-2: android
