+++ device_doctor/lib/src/list_processes.dart
@@ -8,6 +8,7 @@ import 'dart:io';
 import 'package:meta/meta.dart';
 import 'package:path/path.dart' as path;
 
+import 'process_helper.dart';
 import 'utils.dart';
 
 /// Returns a list of Process IDs (PIDs).
